Add a separate LootUtils class
